FIX pylint + Update/complete doc
