+++ src/openhems/modules/energy_strategy/energy_strategy.py
@@ -2,7 +2,8 @@
 Super class for all EnergyStrategy modules
 
 List of todo list to integrate a strategy
-- Implemented : Implement main functions inherit from EnergyStrategy: at least __init__() and updateNetwork() (or eval and apply)
+- Implemented : Implement main functions inherit from EnergyStrategy:
+   at least __init__() and updateNetwork() (or eval and apply)
 - Call : Call constructor in server.py
 - Conf : Configure it in openhems_default.py
 - TestAuto : Add unittest in test_xxx_strategy.py.
@@ -251,3 +252,22 @@ class EnergyStrategy:
 		self.check(now)
 		self.apply(cycleDuration, now=now)
 		return cycleDuration
+
+
+	def switchOnMax(self, cycleDuration):
+		"""
+		Switch on nodes, but 
+		 - If there is no margin to switch on, do nothing.
+		 - Only one (To be sure to not switch on to much devices)
+		"""
+		self.logger.info("%s.switchOnMax()", self.strategyId)
+		marginPower = self.network.getMarginPowerOn()
+		if marginPower<=0:
+			self.logger.info("Can't switch on devices: not enough power margin : %s", marginPower)
+			return True
+		for elem in self.getNodes(True):
+			switchOn = self.switchSchedulable(elem.node, cycleDuration, True)
+			if switchOn and elem.changed(switchOn):
+				self.logger.info("Switch on just one device at each loop to ensure Network constraint.")
+				return True
+		return False
